requirements.txt
dlib==19.24.0
gdown==4.5.1
imageio==2.19.3
keras==2.9.0
keras_ocr==0.9.1
matplotlib==3.5.2
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.2.0
scikit-image==0.19.3
scipy==1.8.1
streamlit==1.10.0
tensorflow==2.9.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0